refactor: CLOVA API 호출 시 전달할 JSON 객체 생성 로직 분리
